Fix size_t/int format specifier mismatch in initializeFrameRates log statements
Agent-Logs-Url: https://github.com/rdkcentral/devicesettings/sessions/2a1df342-47cc-4d33-85c0-b0963487d4a5

+++ ds/frameRate.cpp
@@ -159,7 +159,7 @@ namespace device {
             #endif
                 default: {
                     framerateInfo.name = "Unknown";
-                    INT_WARN("Invalid frame rate id: %d", i);
+                    INT_WARN("Invalid frame rate id: %d", static_cast<int>(i));
                 }
                 break;
             }
@@ -183,11 +183,11 @@ namespace device {
                     framerateInfo.name = tempFloatName;
                 }
                 else {
-                    INT_WARN("Failed to convert frame rate value: %f to string for frame rate id: %d", framerateInfo.value, i);
+                    INT_WARN("Failed to convert frame rate value: %f to string for frame rate id: %d", framerateInfo.value, static_cast<int>(i));
                 }
             }
 
-            INT_INFO("Frame rate id: %d, name: %s, value: %f", i, framerateInfo.name.c_str(), framerateInfo.value);
+            INT_INFO("Frame rate id: %d, name: %s, value: %f", static_cast<int>(i), framerateInfo.name.c_str(), framerateInfo.value);
             // inserting the frame rate info in the map with frame rate id as key and FrameRateInfo struct as value
             _frameRates.insert({static_cast<dsVideoFrameRate_t>(i), framerateInfo});
         }
